+++ src/components/header/Header.scss
@@ -17,7 +17,6 @@ header {
       }
 
       span {
-        cursor: pointer;
         margin: 0 calc(var(--base-space) * 2);
       }
