+++ src/utils/Router.tsx
@@ -1,5 +1,5 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import Main from "../pages/main/Main";
+import Main from "../pages/Main";
 export default function Router() {
   return (
     <>
